ci: Chicken and the egg.
When doing dockerimage changes we want to build the package using this
build environment. Two things are needed to accomplish this.

First; rules in the CI control file to use the appropiate image.

Secondly; the runner must have the pull-policy set to 'if-not-present'.

The latter however requires that we cleanup the server registry
occasionally of remote images.

Note that until we have sticky runners / workspace (gitlab-org/gitlab-ce#29447)
this is the best we can do.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>